requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.7.3
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
Pillow==10.2.0
Requests==2.31.0
streamlit==1.31.1
streamlit_lottie==0.0.5
